Building type declarations from sources
