+++ AGENTS.md
@@ -5,7 +5,7 @@ This file provides guidance to coding agents when working with code in this repo
 ## Commands
 
 - `npm start` — Vite dev server for the docs/demo page (`src/index.html`) at http://127.0.0.1:5173
-- `npm run build` — builds docs (`docs/`) plus both library bundles (`dist/`) in parallel
+- `npm run build` — builds docs (`docs/`) in parallel with the library bundles and type declarations (`dist/`)
 - `npm run lint` — everything: eslint (JS/TS/JSON/MD/CSS), stylelint, `tsc --noEmit`, `attw --pack`
 - `npm test` — Playwright (chromium, firefox, webkit) with nyc/istanbul coverage; auto-starts the dev server
 - `npm run playwright-interactive` — Playwright UI mode
@@ -24,12 +24,22 @@ Vanilla-TS lightbox with two published entry points built from one core:
 - `src/lib/container.ts` — the container is a single `<dialog>` created at import time and shown with `showModal()`. Its `cancel` event is suppressed so Escape is handled by the lightbox's own keyboard navigation rather than the browser. Because the dialog is modal, the page behind it is inert — tests use `clickThroughModal` in `tests/helpers.ts` to drive app buttons while it is open.
 - `src/lib/imagelightbox.jquery.ts` — thin legacy adapter mapping the core onto `$.fn.imageLightbox` and re-emitting the `ilb:*` events as jQuery events. Behaviour changes belong in the core, not here.
 - Events (`ilb:start`, `ilb:quit`, `ilb:loaded`, `ilb:previous`, `ilb:next`, `ilb:error`) are plain bubbling DOM events dispatched from the container or the anchor.
-- CSS lives next to the module that injects it (`arrows.css` beside `arrows.ts`) and is imported from the TS; `src/index.d.ts` declares `*.css`.
-- Public types are **hand-written** in `types/` (both `.d.ts` and `.d.cts` per entry point) and are not generated by the build. `ILBOptions` is a global ambient interface declared there, so adding an option means editing all four type files plus the README option table.
+- CSS lives next to the module that injects it (`arrows.css` beside `arrows.ts`) and is imported from the TS; `src/lib/css.d.ts` declares `*.css` (it must stay inside `src/lib`, the `entryRoot` of the declaration build).
+- Public types are **generated** into `dist/` by `unplugin-dts` (api-extractor rollup), so adding an option means editing `src/lib/interfaces/ILBOptions.ts` and the README option table — nothing else. `ILBOptions` is an ordinary exported interface, re-exported from `src/lib/imagelightbox.ts`. The `.d.cts` variants are plain copies made by `build:lib:cjs-types`.
+- Genuine global augmentations (`Document` for the `ilb:error` overload, `JQuery` for the jQuery methods) live in `declare global` blocks **inside their entry module** — api-extractor only carries an ambient block into the rollup if it is in the entry file itself.
 
 ### Builds
 
 Three separate Vite configs, all with `root: "src"`: `native.vite.config.ts` and `jquery.vite.config.ts` produce the ESM + UMD library bundles in `dist/` (jQuery external), `docs.vite.config.ts` builds the demo page into `docs/` and adds `vite-plugin-istanbul` so Playwright can collect coverage.
+
+Both lib configs also run `unplugin-dts`, which makes their ordering load-bearing — `build:lib` uses `run-s`, not `run-p`:
+
+- The jQuery build runs **first**. Its declaration pass emits `imagelightbox.d.ts` as an intermediate and then deletes it, which would wipe the native build's finished output if the native build ran earlier or concurrently. `build:lib:jquery` renames its own rollup afterwards, because the plugin takes the type-entry filename from the top-level `"types"` field (which has to keep pointing at the native entry so node10 resolution works).
+- The native config needs `exclude: ["src/lib/imagelightbox.jquery.ts"]`, otherwise the `JQuery` augmentation leaks into the native rollup and non-jQuery consumers fail with `TS2315: Type 'JQuery' is not generic`.
+
+`npm run lint`'s `attw --pack` covers all of this: it resolves both entry points under node10, node16 (CJS and ESM) and bundler, so a broken declaration or a stale `types`/`style` pointer fails the lint job.
+
+Only `dist/`, `jquery/` and the README ship to npm — `docs/` is built for GitHub Pages but excluded from `files`, since the demo images and video are several megabytes.
 
 ### Tests
