+++ native.vite.config.ts
@@ -1,5 +1,6 @@
 import { resolve } from "path";
 import webpackStats from "rollup-plugin-webpack-stats";
+import dts from "unplugin-dts/vite";
 import { defineConfig } from "vite";
 
 export default defineConfig({
@@ -17,7 +18,17 @@ export default defineConfig({
     },
     sourcemap: true,
   },
-  plugins: [webpackStats({ fileName: "webpack-stats.native.json" })],
+  plugins: [
+    dts({
+      bundleTypes: true,
+      entryRoot: "src/lib",
+      exclude: ["src/lib/imagelightbox.jquery.ts"],
+      include: ["src/lib/**/*.ts"],
+      outDirs: "dist",
+      root: __dirname,
+    }),
+    webpackStats({ fileName: "webpack-stats.native.json" }),
+  ],
   publicDir: false,
   root: "src",
 });
