+++ src/fixtures.ts
@@ -5,7 +5,7 @@
  */
 
 import { ImageLightbox } from "./lib/imagelightbox";
-import "./index.d";
+import "./lib/css.d";
 
 declare const TEST: boolean | undefined;
 const isTest: boolean = typeof TEST === "undefined" ? false : TEST;
