+++ src/index.ts
@@ -1,5 +1,5 @@
 import { ImageLightbox } from "./lib/imagelightbox";
-import "./index.d";
+import "./lib/css.d";
 
 declare const TEST: boolean | undefined;
 const isTest: boolean = typeof TEST === "undefined" ? false : TEST;
